implement onX() methods throughout.
This is a breaking change.

These are the preferred methods to override (i.e. onDoTask() instead of doTask() ), if you do not want to be responsible for event dispatch. Overriding only the onX() method will ensure the event is dispatched at the 'correct' point in the timeline, accounting for class inheritance.

Get rid of the ArrayBufferEvent and StringBufferEvent. Added new event types such as CreateHtmlEvent, etc.